Refactored to build packages
