+++ .gitignore
@@ -7,3 +7,6 @@ node_modules
 server/static
 __pycache__
 Pipfile.lock
+/build
+/dist
+*egg-info
