+++ Pipfile
@@ -16,9 +16,10 @@ setuptools = "*"
 wheel = "*"
 
 [scripts]
-unittests = "python -m unittest discover -s ./src/atlas-scientific-web"
-lint = "pylint src/atlas-scientific-web/run_local.py"
-hosted = "python src/atlas-scientific-web/run_hosted.py"
+unittests = "python -m unittest discover -s ./src/"
+lint = "pylint src/atlas_scientific_web/run_local.py"
+serve_dev = "python -m src.atlas_scientific_web.serve_dev"
+serve_prod = "waitress-serve --listen=localhost:8080 --call 'src.atlas_scientific_web:create_app'"
 package = "python setup.py sdist bdist_wheel"
 
 [requires]
