--- src/atlas-scientific-web/api/__init__.py
@@ -1,1 +1,0 @@
-from .api import create_app
