--- src/atlas-scientific-web/run_hosted.py
@@ -1,3 +1,0 @@
-#!flask/bin/python
-import api
-api.create_app().run(host="0.0.0.0", debug = True)
