+++ src/atlas_scientific_web/hardware/capabilities.py
