+++ src/atlas_scientific_web/hardware/models.py
