+++ src/atlas_scientific_web/static/index.html
