+++ src/atlas_scientific_web/static/main.css
