+++ src/atlas_scientific_web_tests/test_co2_device.py
@@ -1,11 +1,12 @@
 import unittest
 from unittest.mock import Mock, call, patch
-
-from atlas_scientific.device import AtlasScientificDeviceBus
 from datetime import datetime, timezone
-import api
 
-from i2c import I2CBusIo
+from atlas_scientific_web.hardware.device import AtlasScientificDeviceBus
+from atlas_scientific_web.hardware.i2c import I2CBusIo
+from atlas_scientific_web.api import create_app
+
+date_time_patch = 'atlas_scientific_web.hardware.device.get_datetime_now'
 
 class Co2DeviceTests(unittest.TestCase):
  
@@ -16,13 +17,13 @@ class Co2DeviceTests(unittest.TestCase):
         self.i2cbus.write = Mock()
         self.i2cbus.ping = Mock() 
         
-        self.app = api.create_app(self.i2cbus).test_client()
+        self.app = create_app(self.i2cbus).test_client()
  
 
     # Sample tests
 
     @patch('time.sleep', return_value=None)
-    @patch('atlas_scientific.device.get_datetime_now', return_value = datetime.fromtimestamp(1582672093, timezone.utc))
+    @patch(date_time_patch, return_value = datetime.fromtimestamp(1582672093, timezone.utc))
     def test_can_sample_atlas_scientific_co2_device_with_ppm_output_enabled(self, datetime_now_mock, patched_time_sleep):
 
         # Arrange
@@ -63,7 +64,7 @@ class Co2DeviceTests(unittest.TestCase):
         self.assertEqual(b'[{"symbol": "ppm", "timestamp": "2020-02-25 23:08:13+00:00", "value": "800", "value_type": "int", "unit_code": "PPM"}]\n', response.data)
 
     @patch('time.sleep', return_value=None)
-    @patch('atlas_scientific.device.get_datetime_now', return_value = datetime.fromtimestamp(1582672093, timezone.utc))
+    @patch(date_time_patch, return_value = datetime.fromtimestamp(1582672093, timezone.utc))
     def test_can_sample_atlas_scientific_co2_device_with_device_temperature_output_enabled(self, datetime_now_mock, patched_time_sleep):
 
         # Arrange
