+++ src/atlas_scientific_web_tests/test_rtd_device.py
@@ -1,11 +1,12 @@
 import unittest
 from unittest.mock import Mock, call, patch
-
-from atlas_scientific.device import AtlasScientificDeviceBus
 from datetime import datetime, timezone
-import api
 
-from i2c import I2CBusIo
+from atlas_scientific_web.hardware.device import AtlasScientificDeviceBus
+from atlas_scientific_web.hardware.i2c import I2CBusIo
+from atlas_scientific_web.api import create_app
+
+date_time_patch = 'atlas_scientific_web.hardware.device.get_datetime_now'
 
 class RtdDeviceTests(unittest.TestCase):
  
@@ -16,13 +17,13 @@ class RtdDeviceTests(unittest.TestCase):
         self.i2cbus.write = Mock()
         self.i2cbus.ping = Mock() 
         
-        self.app = api.create_app(self.i2cbus).test_client()
+        self.app = create_app(self.i2cbus).test_client()
  
 
     # Sample tests
 
     @patch('time.sleep', return_value=None)
-    @patch('atlas_scientific.device.get_datetime_now', return_value = datetime.fromtimestamp(1582672093, timezone.utc))
+    @patch(date_time_patch, return_value = datetime.fromtimestamp(1582672093, timezone.utc))
     def test_can_sample_atlas_scientific_rtd_device(self, datetime_now_mock, patched_time_sleep):
 
         # Arrange
